refactor(tauri/ui): consolidate service calls into less files
